% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testnewnew-package.R
\docType{package}
\name{testnewnew-package}
\alias{testnewnew}
\alias{testnewnew-package}
\title{testnewnew: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JieYinStat/testnewnew}
  \item \url{https://jieyinstat.github.io/testnewnew/}
  \item Report bugs at \url{https://github.com/JieYinStat/testnewnew/issues}
}

}
\author{
\strong{Maintainer}: Jie Yin \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
